Create a separate product flavor to publish on F-Droid
