+++ app/src/fdroid/kotlin/xyz/malkki/neostumbler/location/LocationSourceProvider.kt
@@ -1,0 +1,10 @@
+package xyz.malkki.neostumbler.location
+
+import android.content.Context
+
+
+class LocationSourceProvider(private val context: Context) {
+    fun getLocationSource(): LocationSource {
+        return PlatformLocationSource(context)
+    }
+}
