+++ app/src/full/AndroidManifest.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <uses-permission android:name="android.permission.ACCESS_BACKGROUND_LOCATION" tools:node="merge" />
+    <uses-permission android:name="android.permission.ACTIVITY_RECOGNITION" tools:node="merge"/>
+
+    <application tools:node="merge">
+
+        <receiver
+            android:name=".scanner.autoscan.LocationReceiver"
+            android:enabled="true"
+            android:exported="true"
+            tools:node="merge" />
+
+        <receiver
+            android:name=".scanner.autoscan.RescheduleReceiver"
+            android:enabled="true"
+            android:exported="true"
+            tools:node="merge">
+            <intent-filter>
+                <action android:name="android.intent.action.MY_PACKAGE_REPLACED" />
+            </intent-filter>
+            <intent-filter>
+                <action android:name="android.intent.action.BOOT_COMPLETED" />
+            </intent-filter>
+        </receiver>
+
+        <receiver
+            android:name=".scanner.autoscan.ActivityTransitionReceiver"
+            android:enabled="true"
+            android:exported="true"
+            tools:node="merge" />
+    </application>
+</manifest>
