+++ app/src/full/kotlin/xyz/malkki/neostumbler/scanner/autoscan/ActivityTransitionReceiver.kt
